requirements.txt

keras
tensorflow

gensim
wordsegment
nltk

python-dotenv
tweepy
google-cloud-storage

#h5py
#pandas
#pillow
#numpy
